requirements.txt
autopep8==1.3.2
boto3==1.4.4
coverage==4.1
et3==1.5
git+https://github.com/elifesciences/elife-tools.git@c176a028e7f48455dc8bb5cd409401923769d748#egg=elifetools
Flask==0.12.2
Flask-Testing==0.6.2
future==0.16.0
green==2.9.0
isbnlib==3.7.2
joblib==0.11
jsonschema==2.6.0
mock==2.0.0
newrelic==2.88.0.72
pyflakes==1.5.0
pylint==1.7.2
python-dateutil==2.6.0
python-json-logger==0.1.7
python-slugify==1.2.4
pytz==2017.2
PyYAML==3.12
requests==2.18.1
requests-aws==0.1.8
requests-cache==0.4.13
rfc3339==5.2
strict-rfc3339==0.7
uWSGI==2.0.15

#connexion==1.1.10
git+https://github.com/elifesciences/connexion.git@995ebab9a5731c664434f0fc8f8fd1bc1acfa5b0#egg=connexion
flex==6.10.0
